lint and code-climate fix
